reference-implementations/validator-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-protocol-validate"
version = "0.1.0"
description = "Language-native reference validator for the agent-protocol Change Manifest"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
    "jsonschema>=4.17",
]

[project.scripts]
agent-protocol-validate = "agent_protocol_validate.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
